factor out biquad, lowpass, highpass to sox compatibility
